Cleanup
Felt silly that WritableImage had a ColorModel() method, so I got rid
of it. That means I had to replace the WritableImage in SwitchyRGBA
with just an RGBA, which implements ColorModel(), so it's an Image.
Seems clearer to me, though might actually be worse in some abstract OO
sense.

- Add a usage string and actually check some error codes.
- Add a .gitignore so I don't have to actually bother with a directory
structure yet